[Refactor] Channel Dto
