+++ src/service/nestjs/src/api/channel/channel.service.ts
@@ -16,7 +16,7 @@ class ChannelService {
 		}
 	}
 
-	async createChannel(channelRequestDto: Dto.Request.ChannelDto): Promise<ChannelModel> {
+	async createChannel(channelRequestDto: Dto.Request.Channel): Promise<ChannelModel> {
 		try {
 			const title: string = channelRequestDto.title;
 			const visibility: ChannelVisibility = channelRequestDto.visibility;
@@ -30,7 +30,7 @@ class ChannelService {
 		}
 	}
 
-	async updateChannel(id: string, updateChannelDto: Dto.Request.Channel.Update.PartialChannel) {
+	async updateChannel(id: string, updateChannelDto: Partial<Dto.Request.Channel>) {
 		try {
 			return await this.prismaService.channel.update({
 				where: { id },
