+++ src/service/nestjs/src/api/channel/dto/index.ts
@@ -1,1 +1,2 @@
 export * as Request from './request';
+export * as Response from './response';
