--- src/service/nestjs/src/api/channel/dto/request/channel/update/partial-channel.dto.ts
@@ -1,4 +1,0 @@
-import { PartialType } from '@nestjs/swagger';
-import { Channel } from '../channel.dto';
-
-export class PartialChannel extends PartialType(Channel) {}
