+++ src/service/nestjs/src/api/channel/dto/request/index.ts
@@ -1,2 +1,1 @@
-export * as Channel from './channel';
-export { Channel as ChannelDto } from './channel/channel.dto';
+export * from './channel';
