+++ src/service/nestjs/src/api/channel/dto/response/index.ts
@@ -1,0 +1,1 @@
+export * from './participant';
